save: SC-block offsets; field-VM nD sub-4/5 (VRAM STP); mdec STR header layout (#81)
* save: pin story-flags + inventory SC-block offsets; field-VM: nD sub-4/5 + cleanup

crates/save/src/card.rs:
- Add RETAIL_STORY_FLAGS_OFFSET (0x14C0), RETAIL_STORY_FLAGS_SIZE (0x200)
- Add RETAIL_INVENTORY_OFFSET (0x1818), RETAIL_INVENTORY_SLOTS (72),
  RETAIL_INVENTORY_SIZE (0x90)
- Add SAVE_GAME_DATA_RAM_BASE (0x80084340) with derivation comments
- Add read_retail_story_flags() / read_retail_inventory() slice helpers
- Re-export all new constants + helpers from crate root

Offsets verified by byte-matching mednafen main-RAM dumps (mc0-mc9 save
states) against real MCR save blocks for both Drake and Sebucus saves.
100% match for inventory; formula: block_offset = 0x200 + (ram - 0x80084340).

crates/engine-vm/src/field.rs — 0x4C nibble D sub-4/5:
- Port 0x4C nD sub-4 (34-byte u16×16 OR mask) and sub-5 (AND mask);
  host hooks op4c_n_d_sub_4_or_mask / op4c_n_d_sub_5_and_mask (default no-op)
- Coverage matrix row D now entirely ✓ (was P for sub-4/5)
- Fix nibble-C dead _ => Pending arm → unreachable! (all 16 sub-ops covered)
- Fix nibble-E _ => Pending → _ => Halt (sub-F falls through to default halt)
- Update stale preamble comments for nibbles 8, 9, D, E to reflect current state
- Add 3 tests: nD sub-4 advances 34 bytes, nD sub-5 AND semantics, nE sub-F halts

docs: add complete SC block layout table with confirmed offsets; update
script-vm.md coverage matrix row D and remove resolved blocker entry.
scripts/find_save_offsets.py: analysis script used to confirm the offsets.


* field-VM: nD sub-4/5 are 6-byte VRAM STP set/clear, not 34-byte mask

The earlier interpretation of `0x4C nD sub-4/5` as 34-byte instructions
carrying 16 inline u16 mask words was wrong on all three axes (length,
operands, semantics). Raw MIPS at `0x801e2c44..0x801e2d60` shows two
`func_0x8003ce9c` u16 reads (4 operand bytes) and `addiu s8, s8, 0x6`
(6-byte PC advance). The 16-element buffer is stack-local; it's pixels
read from VRAM at runtime via the PsyQ libgs sequence
`DrawSync; StoreImage(rect, buf); DrawSync; <fold>; LoadImage(rect, buf)`.

- sub-4 (`ori v0, v0, 0x8000`): set STP bit on non-zero pixels
- sub-5 (`andi v0, v0, 0x7fff` guarded by `bne v0, a1 (=0x8000)`):
  clear STP bit unless pixel is already 0x8000

Host hooks renamed to `op4c_n_d_sub_4_vram_stp_set(x, y)` /
`op4c_n_d_sub_5_vram_stp_clear(x, y)`. PC advance corrected to 6.
Unit tests rewritten against the 6-byte encoding. Comment block,
script-vm.md sub-cluster section, and find_save_offsets.py
`INVENTORY_END` typo updated.

Also fixes `scripts/find_save_offsets.py:45` where `INVENTORY_END`
was `0x80085A48` (240 bytes) with a comment claiming 0x90 / 144;
the inventory is 144 bytes ending at `0x800859E8`.


* mdec: fix STR sector header layout (canonical PSX-SPX, 32-byte header)

The previous layout had every field after offset 2 shifted by a few bytes
relative to the canonical PSX-SPX STR sector header, and treated `frame_number`
as u16 instead of u32. With those wrong offsets the test
`mv_str_files_decode_first_frame_via_engine_pipeline` would read `frame_size`
as `width` (e.g. 15820 px) and fail the multiple-of-16 dimension check in
`MdecDecoder::decode_frame`, never assembling a usable first frame.

Direct inspection of MV1.STR sectors on disc (LBA 138778+):

  +00  u16  magic (0x0160)
  +02  u16  type  (0x8001 = PSX picture)
  +04  u16  chunk_number
  +06  u16  chunks_per_frame
  +08  u32  frame_number
  +0C  u32  frame_size_bytes
  +10  u16  width  (320)
  +12  u16  height (224)
  +14  16   reserved/picture-code (per-frame constants; not decoder-relevant)
  +20  2016 bs_data

Updates:

- `parse_video_sector` rebuilt against the canonical PSX-SPX offsets.
- `StrSectorHeader.frame_number` is now u32; the unused `bs_version` /
  `quantize_scale` fields are removed (they were at +0x10/+0x12 in the
  old layout, which is actually where width/height live; the real BS
  header lives at the start of the assembled bitstream).
- BS payload start moved to +0x20; `SECTOR_PAYLOAD_BYTES` becomes 2016.
- `StrFrameAssembler.current_frame` typed u32 to match.
- Synthetic-sector test fixture + mdec CLI + engine-shell consumers updated.

Verified end-to-end against the on-disc MV*.STR files: all 6 files
(MV1..MV6) now decode their first frame at 320x224 (frame#1).


---------

+++ crates/engine-shell/tests/fmv_table_roundtrip.rs
@@ -62,7 +62,7 @@ fn decode_first_frame(sectors: &[u8]) -> Option<VideoFrame> {
                         rgba,
                         width: hdr.width as u32,
                         height: hdr.height as u32,
-                        frame_number: hdr.frame_number as u32,
+                        frame_number: hdr.frame_number,
                     });
                 }
             }
